requirements (1).txt
streamlit
tflite-runtime==2.10.0
numpy
pillow
